streamlit_app/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.7.0
plotly>=5.15.0
opencv-python>=4.8.0
